feat(mini-files): Add mini.files plugin and remove oil.nvim
This commit introduces the mini.files plugin to the Neovim configuration,
which provides a minimalistic file explorer. The plugin is configured with
a lazy style key map.

In addition, the oil.nvim plugin is removed from the configuration. This
change is reflected in the deletion of the corresponding lua file.