feat(grapple): change scope and keybindings
The scope of the grapple plugin has been changed from 'git' to
'git_branch'. This change will allow more specific control over the
plugin's functionality.

Additionally, the keybindings for selecting tags have been updated.
Instead of using ';a', ';s', ';q', and ';w', the new keybindings are
'<c-a>', '<c-s>', '<c-q>', and '<c-w>'. This change will make the
keybindings more intuitive and easier to use.

A new file 'lazy-lock.json' has also been added to the .gitignore file.
This will prevent unnecessary tracking of this file in the git history.