feat: Add cyberdream.nvim and update .gitignore
This commit does the following:
- Adds the cyberdream.nvim plugin to the lua/plugins/init.lua file.
- Removes the 'spell' entry from the .gitignore file.
- Adds new words to the spell/en.utf-8.add file.
- Creates a new spell/en.utf-8.add.spl file.

The cyberdream.nvim plugin is added with a configuration that includes
transparent, italic_comments, borderless_telescope, and terminal_colors
options. The colorscheme is set to cyberdream.

The 'spell' entry is removed from the .gitignore file, indicating that
files in the 'spell' directory should now be tracked by git.

The spell/en.utf-8.add file is updated with new words that are specific
to Neovim and its configuration.

A new spell/en.utf-8.add.spl file is created, but the contents are not
shown in the diff.